Added non-async functions

+++ commanderbot/core/application_emoji_manager.py
@@ -1,4 +1,4 @@
-from typing import Iterable, Optional
+from typing import Optional
 
 from discord import Attachment, Emoji
 from discord.ext.commands import Bot
@@ -16,7 +16,7 @@ class ApplicationEmojiManager:
         emojis = await self._bot.fetch_application_emojis()
         self._cache = {e.name: e for e in emojis}
 
-    def _find_in_cache(self, emoji: str | EmojiID) -> Emoji:
+    def _get_from_cache(self, emoji: str | EmojiID) -> Optional[Emoji]:
         # Try to find the emoji by its name
         if isinstance(emoji, str):
             if e := self._cache.get(emoji):
@@ -28,7 +28,8 @@ class ApplicationEmojiManager:
                 if e.id == emoji:
                     return e
 
-        raise ApplicationEmojiDoesNotExist(emoji)
+    async def build_cache(self):
+        await self._update_cache()
 
     async def fetch(self, emoji: str | EmojiID) -> Emoji:
         """
@@ -51,9 +52,12 @@ class ApplicationEmojiManager:
 
         if not self._cache:
             await self._update_cache()
-        return self._find_in_cache(emoji)
 
-    async def fetch_all(self) -> Iterable[Emoji]:
+        if found_emoji := self._get_from_cache(emoji):
+            return found_emoji
+        raise ApplicationEmojiDoesNotExist(emoji)
+
+    async def fetch_all(self) -> list[Emoji]:
         """
         Fetch all application emojis from Discord, or from the cache.
 
@@ -67,7 +71,24 @@ class ApplicationEmojiManager:
 
         if not self._cache:
             await self._update_cache()
-        return self._cache.values()
+        return list(self._cache.values())
+
+    def get(self, emoji: str | EmojiID) -> Optional[Emoji]:
+        """
+        Get an application emoji from the cache. It may or may not exist.
+
+        Parameters
+        ----------
+        emoji: :class:`str | EmojiID`
+            The emoji to get.
+        """
+        return self._get_from_cache(emoji)
+
+    def get_all(self) -> list[Emoji]:
+        """
+        Get all application emojis from the cache. They may or may not exist.
+        """
+        return list(self._cache.values())
 
     async def create(self, name: str, image: bytes | Attachment) -> Emoji:
         """
@@ -125,10 +146,11 @@ class ApplicationEmojiManager:
             await self._update_cache()
 
         # Edit the emoji
-        found_emoji: Emoji = self._find_in_cache(emoji)
-        renamed_emoji: Emoji = await found_emoji.edit(name=new_name)
-        await self._update_cache()
-        return renamed_emoji
+        if found_emoji := self._get_from_cache(emoji):
+            edited_emoji: Emoji = await found_emoji.edit(name=new_name)
+            await self._update_cache()
+            return edited_emoji
+        raise ApplicationEmojiDoesNotExist(emoji)
 
     async def delete(self, emoji: str | EmojiID):
         """
@@ -153,6 +175,8 @@ class ApplicationEmojiManager:
             await self._update_cache()
 
         # Delete the emoji
-        found_emoji: Emoji = self._find_in_cache(emoji)
-        await found_emoji.delete()
-        await self._update_cache()
+        if found_emoji := self._get_from_cache(emoji):
+            await found_emoji.delete()
+            await self._update_cache()
+            return
+        raise ApplicationEmojiDoesNotExist(emoji)
